Consolidate geometry headers
