+++ src/platforms/x11/input/input_device.h
@@ -20,8 +20,8 @@
 #include "mir/input/input_device.h"
 #include "mir/input/input_device_info.h"
 #include "mir_toolkit/event.h"
-#include "mir/geometry/point_f.h"
-#include "mir/geometry/displacement_f.h"
+#include "mir/geometry/point.h"
+#include "mir/geometry/displacement.h"
 #include "mir/optional_value.h"
 
 #include <chrono>
